% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_fit.R
\name{combine.s2bak}
\alias{combine.s2bak}
\title{Combine SO, S2 and BaK objects}
\usage{
combine.s2bak(so = NULL, s2 = NULL, bak = NULL)
}
\arguments{
\item{so}{output from fit.s2bak.so or fit.s2bak.s2 without survey data.}

\item{s2}{output from fit.s2bak.s2 function}

\item{bak}{output from fit.s2bak.bak function}
}
\value{
Object of class s2bak, equivalent to having run
\link[s2bak]{fit.s2bak}
}
\description{
Combines separately fitted SO, S2 and BaK models into a single
S2BaK object. The output will be identical to running the entire process in
fit.s2bak function, and can be used in the same situations.

Models can be partially provided, for instance only SO and BaK, in which case
the use of s2bak predict functions will not apply S2 but instead only run
predictions with sightings-only and BaK adjustment.
}
