% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_class.R
\name{display_bak}
\alias{display_bak}
\title{Display information for s2bak.bak class object}
\usage{
display_bak(object)
}
\arguments{
\item{object}{object of class \code{s2bak.bak}}
}
\value{
Printout of information
}
\description{
Display information for s2bak.bak class object
}
