% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_class.R
\name{display_sos2}
\alias{display_sos2}
\title{Display information for s2bak.so and s2bak.s2 class object}
\usage{
display_sos2(object)
}
\arguments{
\item{object}{object of class \code{s2bak.so} or \code{s2bak.s2}}
}
\value{
Printout of information
}
\description{
Display information for s2bak.so and s2bak.s2 class object
}
