% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_etc.R
\name{dst}
\alias{dst}
\title{Distance function}
\usage{
dst(x, y)
}
\arguments{
\item{x}{Numerical first value, order does not matter}

\item{y}{Numerical second value, order does not matter}
}
\value{
Square root of the sum of the values squared
}
\description{
Calculate distance between two values
}
