% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_predict.R
\name{predict.s2bak}
\alias{predict.s2bak}
\alias{predict.s2bak.s2}
\alias{predict.s2bak.so}
\alias{predict.s2bak.bak}
\title{Make predictions using fitted SO, S2, BaK or S2BaK class models

#' @description S3 methods for making models predictions using the output
from \link[s2bak]{fit.s2bak}, \link[s2bak]{fit.s2bak.so},
\link[s2bak]{fit.s2bak.s2} and \link[s2bak]{fit.s2bak.bak}.

If the complete model of class \code{s2bak} is provided, the user can specify
what type of model predictions to use to allow for specific outputs from
the sub-components of the model. When applying \code{predict.s2bak},
the function assumes that \code{predict.fun} is the same for all models.}
\usage{
\method{predict}{s2bak}(
  model,
  newdata,
  trait = NA,
  predict.fun,
  predict.bak.fun,
  output = c("s2bak", "all", "so", "s2", "sobak")[1],
  ncores = 1,
  useReadout = FALSE,
  ...
)

\method{predict}{s2bak.s2}(model, newdata, predict.fun, useReadout = FALSE, ncores = 1, ...)

\method{predict}{s2bak.so}(model, newdata, predict.fun, useReadout = FALSE, ncores = 1, ...)

\method{predict}{s2bak.bak}(
  model,
  predictions,
  trait,
  data,
  predict.bak.fun,
  truncate = c(1e-04, 0.9999)
)
}
\arguments{
\item{model}{Output from fit.s2bak.bak(), with fitted BaK model}

\item{newdata}{A data.frame containing the values . All variables needed for
prediction should be included.}

\item{trait}{Trait data, with column 'species' matching those in predictions.}

\item{predict.fun}{Predict function linked to the SDM used. Functions have
the structure of model and newdata as the first and second arguments,
respectively.}

\item{predict.bak.fun}{Model function for predicting bias adjustment model
(e.g., \link[stats]{predict.glm}). Needs to match \code{bak.fun}}

\item{output}{The choice of how predictions are made using the \code{s2bak} model.

Only one type of output can be selected: sightings-only (output = "so"),
sightings-survey (S2; output = "s2") and bias-adjusted sightings-only
(output = "sobak").

If interested in using the \code{s2bak.bak} model on its own with provided model
predictions, see \link[s2bak]{predict.s2bak.bak}.

The default method is \code{s2bak}, which will return a single set
of predictions for each species, prioritizing
S2 models over bias-adjusted sightings-only models and .
If the full S2BaK model is provided, the function will
make predictions using \link[s2bak]{s2bak.predict.s2} for species with
sightings and survey data, and adjusted sightings-only predictions using BaK
for species without survey data.
If only \code{s2bak.so} and \code{s2bak.bak} models are provided
(that is, an S2BaK class model with S2 is NA), the functions returns only
adjusted sightings-only predictions, requiring trait data for the species.
Conversely, if only \code{s2bak.s2} model is provided, then predictions are made
only with s2bak.predict.s2.

#' If the \code{all} is selected, all sets of predictions are returned as a
list: \code{so}, \code{s2}, \code{sobak} and \code{s2bak}.
If models are missing, then only available predictions are made.
For example, if the \code{s2bak} model contains only
\code{s2bak.s2} and \code{s2bak.so} models, then only those model
predictions will be returns without BaK adjustment.}

\item{ncores}{Number of cores to fit the SDMs, default is 1 core but can be
automatically set if ncores=NA. If ncores > number of available cores - 1,
set to the latter.}

\item{useReadout}{logical; if TRUE will do readout over stored SDMs.
If there are no SDMs then it will automatically check for readout}

\item{...}{Additional arguments passed into function for
prediction (predict.fun).}

\item{predictions}{Sightings-only predictions as a matrix or data.frame with
rows as sites and columns as species. Assumes as type="response", and rows of
data.frame correspond to newdata rows.}

\item{data}{Environmental data, with rows corresponding to rows of
predictions}

\item{truncate}{Numeric minimum and maximum range of predicted values. Values
very close to zero or one cannot be meaningfully distinguished, however
these extreme values may have disproportionally large consequences on
likelihoods due to logit transformation.}

\item{survey_var}{Character name for the predictor variable determining
a site is sightings-only (1) or survey data (0).
The column is automatically within the function, and is used to define with
formula.}
}
\value{
Model predictions as a data.frame with columns for each species and
rows for each location. If output = "all" is selected, a lot of
data.frame predictions are returned.

Generates a matrix of predictions with rows being indices in the
data.frame, and columns representing each species.

Model predictions but with adjustments made by the BaK model.
Note the default right now is type="response"
}
\description{
The function predict.s2bak.s2 is used for \code{s2bak.so} and
\code{s2bak.s2} models. It automatically detects which
model class is provided and makes the appropriate adjustments.
\link[s2bak]{predict.s2bak.so} is a wrapper function
that detects the class of the inputed model and makes the appropriate
prediction.

For \code{s2bak.s2} models, the binary variable denoting sightings-only or survey
is stored and will be checked by the function.

predict.s2bak.bak adjusts sightings-only predictions using a
fitted s2bak.bak model, using trait, environmental data and
sightings-only (\link[s2bak]{predict.s2bak.so}) predictions.
}
