% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_fit.R
\name{replace.s2bak}
\alias{replace.s2bak}
\title{Add model to object of class s2bak}
\usage{
replace.s2bak(x, object)
}
\arguments{
\item{x}{object of class \code{s2bak.so}, \code{s2bak.s2} or \code{s2bak.bak}}

\item{object}{object of class \code{s2bak} with fields to be replaced

#' @examples
\subsection{See ?fit.s2bak}{
}}
}
\value{
Object of class s2bak, with the object slot replaced with \code{x}
}
\description{
s2bak objects can be incomplete, for example when using the
\link[s2bak]{combine.s2bak} function without specifying all arguments.
This function allows the field of a given \code{s2bak} object to be replaced
with a new fitted model of class \code{s2bak.so}, \code{s2bak.s2} or \code{s2bak.bak}.
}
