% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_etc.R
\name{s2bak.addPred}
\alias{s2bak.addPred}
\title{Add predictor to the formula.}
\usage{
s2bak.addPred(formula, pred = "so")
}
\arguments{
\item{formula}{Formula or list of formulas to add 'pred' to}

\item{pred}{Predictor name to add to formula}
}
\value{
A formula with added predictor
}
\description{
Adds 'so' (or other specified predictor) to the formula as an
additional predictor. The arugment 'formula' assumes the formula follows a
"Y ~ X" format. The function will not add the predictor if it is already
in the formula.
}
