% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_funcs.R
\name{s2bak.BaK}
\alias{s2bak.BaK}
\title{Build bias adjustment kernel (BaK) for a fitted sightings-only SDM.

\link{DESC}.

NOTE: ASSUMES PREDICTIONS, ENVIRONMENTAL DATA AND SURVEY DATA ROWS ALIGN. MAYBE MAKE FORMAT IT ALIGN WITH s2bak.fit() (that is, allow for indexing).}
\usage{
s2bak.BaK(predictions, surv, data, trait, verbose = FALSE)
}
\arguments{
\item{predictions}{= Sightings-only (SO) model predictions over the survey sites for all species, beyond those found in the survey data, as a matrix with columns for each species}

\item{surv}{= Survey data as a matrix, with rows corresponding to environmental data rows.}

\item{data}{= Survey environmental data. Assumes all columns correspond to relevant environmental data and match with predictions/survey data.}

\item{trait}{= Full trait data, as a data.frame with 'species' as a column and relevant traits for the rest of them. Do not necessarily have to be involved in the survey, but will be used in the final adjustment model as final output.}

\item{verbose}{= Make print statements}
}
\value{
Bias adjustment models, the kernels (location and species), as a second-order GLM.
}
\description{
Build bias adjustment kernel (BaK) for a fitted sightings-only SDM.

\link{DESC}.

NOTE: ASSUMES PREDICTIONS, ENVIRONMENTAL DATA AND SURVEY DATA ROWS ALIGN. MAYBE MAKE FORMAT IT ALIGN WITH s2bak.fit() (that is, allow for indexing).
}
