% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_funcs.R
\name{s2bak.combine}
\alias{s2bak.combine}
\title{Combine separately fitted SO, S2 and BaK models into a single S2BaK object}
\usage{
s2bak.combine(SO = NULL, S2 = NULL, BaK = NULL)
}
\arguments{
\item{SO}{= output from s2bak.SO or s2bak.S2 without survey data.}

\item{S2}{= output from s2bak.S2 function}

\item{BaK}{= output from s2back.BaK function}
}
\value{
Object of class s2bak.S2Bak, equivalent to having run s2bak.S2BaK() for the entire dataset
}
\description{
Assuming everything was done correctly in the previous, the output should mimic an S2BaK function
}
\details{
It is possible to only provide SO and BaK, in which case the use of s2bak predict functions will not apply S2 but instead only run predictions with sightings-only and BaK adjustment.
}
