% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_funcs.R
\name{s2bak.predict.BaK}
\alias{s2bak.predict.BaK}
\title{Adjust SO model predictions using BaK

Make model adjustments using the output from the BaK output, requiring trait, environmental data and SO predictions.}
\usage{
s2bak.predict.BaK(predictions, bak, trait, data)
}
\arguments{
\item{predictions}{= Sightings-only predictions as a matrix or data.frame with rows as sites and columns as species. Assumes as type="response", and rows of data.frame correspond to newdata rows.}

\item{bak}{= Output from s2bak.BaK(), with fitted BaK model}

\item{trait}{= Trait data, with column 'species' matching those in predictions.}

\item{data}{= Environmental data, with rows corresponding to rows of predictions}
}
\value{
Model predictions but with adjustments made by the BaK model. Note the default right now is type="response"
}
\description{
Adjust SO model predictions using BaK

Make model adjustments using the output from the BaK output, requiring trait, environmental data and SO predictions.
}
