% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_funcs.R
\name{s2bak.predict.SOS2}
\alias{s2bak.predict.SOS2}
\title{Make predictions using fitted SO, S2, BaK or S2BaK class models

It should automatically detect which model we have based on output. Can provide either output from s2bak.S2 or s2bak.SO.}
\usage{
s2bak.predict.SOS2(
  model,
  newdata,
  predict.fun = predict.gam,
  doReadout = FALSE,
  verbose = FALSE,
  ncores = 1,
  ...
)
}
\arguments{
\item{model}{= Fitted models to use for prediction. If the object does not have stored SDMs, it will check to see if there is readout (alternatively, you could force readout with doReadout = T).}

\item{newdata}{= A data.frame containing the values . All variables needed for prediction should be included.}

\item{predict.fun}{= Predict function linked to the SDM used. The default used is predict.gam from the mgcv package. Functions have the structure of model and newdata as the first and second arguments, respectively.}

\item{doReadout}{= logical; if TRUE will do readout over stored SDMs. If there are no SDMs then it will automatically check for readout}

\item{verbose}{= If true, provide messages as we go.}

\item{ncores}{= Number of cores to fit the SDMs, default is 1 core but can be automatically set if ncores=NA. If ncores > number of available cores - 1, set to the latter.}

\item{...}{= Additional arguments passed into function for prediction (predict.fun).}

\item{class}{= Whether we're dealing with SO or S2 models, which is only used when providing a directory so we can identify which class we are interested in.}
}
\value{
Generates a matrix of predictions with rows being indices in the data.frame, and columns representing each species.
}
\description{
Make predictions using fitted SO, S2, BaK or S2BaK class models

It should automatically detect which model we have based on output. Can provide either output from s2bak.S2 or s2bak.SO.
}
