% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_funcs.R
\name{s2bak.S2BaK}
\alias{s2bak.S2BaK}
\title{@title Fit entire S2Bak model}
\usage{
s2bak.S2BaK(
  formula,
  data,
  obs,
  surv,
  trait,
  background = NA,
  sdm.fun = gam,
  predict.fun = predict.gam,
  overlapbackground = TRUE,
  nbackground = 10000,
  surv.formula = TRUE,
  ncores = 1,
  verbose = FALSE,
  readout = NA,
  version = c("full", "short")[1],
  ...
)
}
\arguments{
\item{...}{= Same as above, with "..." for the fitting of SO and S2}
}
\value{
An S2BaK class object containing S2, SO and BaK.
}
\description{
Build S2BaK from top to bottom. Has functionality for parallelisation, but the default is 1 core.
}
\details{
Fits SO models for all species, S2 models for species with survey data. \link{Combine with S2 and SO, note that we include predict.fun, traite and force surv as a mandatory input}
Right now, assumes that all columns/variables in 'data' are relevant for the location bias model (minus the index or whatever).
}
