% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_etc.R
\name{s2bak.scale}
\alias{s2bak.scale}
\title{Scales provided columns to a mean of 0 and standard deviation of 1. Does not scale binary and categorical predictors. If providing projected data as well, then it will scale them using the mean and standard deviation of the inputted environment and output both as a list.}
\usage{
s2bak.scale(data, project = NA, getScaleValues = FALSE)
}
\arguments{
\item{data}{Values to scale, with column names indicating the variable}

\item{project}{Projected values to scale, using the mean and standard deviations of 'env'. If NA, then the function will only scale the environmental values provided}

\item{getScaleValues}{Return mean and standard deviations of each factor}
}
\value{
Returns a data.frame of scaled columns for non-categorical and non-binary variables. If getScaleValues or project data is provided, returns as a list.
}
\description{
Scales provided columns to a mean of 0 and standard deviation of 1. Does not scale binary and categorical predictors. If providing projected data as well, then it will scale them using the mean and standard deviation of the inputted environment and output both as a list.
}
