% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_etc.R
\name{s2bak.truncate}
\alias{s2bak.truncate}
\title{Truncate values based on min and max input}
\usage{
s2bak.truncate(x, min = NA, max = NA)
}
\arguments{
\item{x}{Numerical values to truncate}

\item{min}{Minimum value to truncate, where \code{x <= min} with be
\code{min}. If left as NA it will use the minimum value of x.}

\item{max}{Maximum value to truncate, where \code{x >= max} with be
\code{max}. If left as NA it will use the maximum value of x.}
}
\value{
Truncated x values
}
\description{
Truncate provided numerical values based on a defined minimum
and maximum cutoff. Used in \link[s2bak]{fit.bak} and
\link[s2bak]{predict.bak}.
}
