% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_summary.R
\name{summary.s2bak}
\alias{summary.s2bak}
\alias{summary.s2bak.s2}
\alias{summary.s2bak.so}
\alias{summary.s2bak.bak}
\title{Generate summary statistics for s2bak objects}
\usage{
\method{summary}{s2bak.s2}(object, species, ...)

\method{summary}{s2bak.so}(object, species, ...)

\method{summary}{s2bak.bak}(
  object,
  bak = c("all", "bias_site", "bias_species", "bias_adj")[1],
  ...
)

\method{summary}{s2bak}(
  object,
  submodel = c("all", "so", "s2", "bak")[1],
  species = NULL,
  bak = c("all", "bias_site", "bias_species", "bias_adj")[1],
  ...
)
}
\arguments{
\item{object}{Object of class s2bak\verb{, }s2bak.so\verb{, }s2bak.s2\code{or}s2bak.bak`.}

\item{species}{For \code{s2bak.so}, \code{s2bak.s2} and \code{s2bak} models.
The character name of the species to access the model,
or the index if an integer is provided.}

\item{...}{Additional arguments provided by the object's \code{summary} function.}

\item{bak}{For \code{s2bak.bak} and \code{s2bak} models. Select which of the three
submodels to display, or all of them (default bak = "all").}

\item{submodel}{Which submodel (\code{s2bak.so}, \code{s2bak.s2} or \code{s2bak.bak})
to display (default submodel = "all", which displays all of them).
Note that species must be specified if attempting a summary of \code{s2bak.so}
or \code{s2bak.s2} models.}
}
\value{
The relevant summary output for a given species.
}
\description{
This function will allow the user to call the \code{summary}
function for \code{s2bak}, \code{s2bak.so}, \code{s2bak.s2} and \code{s2bak.bak} class objects.

This model assumes that the model used in the fitting process has an
associated \code{summary} function associated with it.

As these models may be fitted for multiple species or sub-models, the user
can specify which index or species to view.
}
