% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_tgb.R
\name{tgb_sample}
\alias{tgb_sample}
\title{Sample target-group background sites}
\usage{
tgb_sample(obs, nbackground = 10000)
}
\arguments{
\item{obs}{Species observations as a data.frame, with a column for species
name (must be labelled 'species') and column of index of observations to
reflect presences. If the index column name is not found in 'data', it
assumes row number.}

\item{nbackground}{The number of background sites to sample. If the number
of sites specified is less than the available sites in \code{data}, then the
maximum will be sampled.}
}
\value{
A vector of indices corresponding to the \code{data} argument, which can
be used to fit SDMs through link[s2bak]{fit.s2bak},
\link[s2bak]{fit.s2bak.so} and \link[s2bak]{fit.s2bak.s2}
}
\description{
This function samples background sites used in species
distribution modelling (SDM), using the target-group background (TGB)
sampling method.

Sites within the species observations, weighted by the number of sightins
in \code{obs}.

without sightings will be excluded from sampling.
}
\examples{
## WIP ##
## But here we would generate data, sample background sites then fit model

}
