% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_predict.R
\name{predict.s2bak.bak}
\alias{predict.s2bak.bak}
\alias{predict.s2bak.s2}
\alias{predict.s2bak.so}
\alias{predict.s2bak}
\title{Make predictions using fitted SO, S2, BaK or S2BaK class models}
\usage{
\method{predict}{s2bak.bak}(predictions, bak, trait, data)

\method{predict}{s2bak.s2}(
  model,
  newdata,
  predict.fun = predict.glm,
  useReadout = FALSE,
  ncores = 1,
  ...
)

\method{predict}{s2bak.so}(
  model,
  newdata,
  predict.fun = predict.glm,
  useReadout = FALSE,
  ncores = 1,
  ...
)

\method{predict}{s2bak}(
  model,
  newdata,
  trait = NA,
  predict.fun = predict.glm,
  ncores = 1,
  useReadout = FALSE,
  ...
)
}
\arguments{
\item{predictions}{Sightings-only predictions as a matrix or data.frame with
rows as sites and columns as species. Assumes as type="response", and rows of
data.frame correspond to newdata rows.}

\item{bak}{Output from fit.s2bak.bak(), with fitted BaK model}

\item{trait}{Trait data, optional if model is S2 or SO}

\item{data}{Environmental data, with rows corresponding to rows of
predictions}

\item{model}{Outputted model from fit.s2bak.so, fit.s2bak.s2, fit.s2bak or
s2bak.combine.}

\item{newdata}{Environmental data to predict}

\item{predict.fun}{prediction function for SDM (for S2 and SO)}

\item{useReadout}{Whether to force useReadout in SOS2}

\item{ncores}{Paralellize SO/S2 predictions, if NA then auto-pick it}

\item{...}{Any additional arguments for the predict.fun}
}
\value{
Model predictions but with adjustments made by the BaK model.
Note the default right now is type="response"

Generates a matrix of predictions with rows being indices in the
data.frame, and columns representing each species.

Model predictions as a data.frame with columns for each species and
rows for each location
}
\description{
When using a fitted s2bak.bak model, adjustments are made using
trait, environmental data and sightings-only (\link[s2bak]{s2bak.so})
predictions.

The function automatically detects which model class is used.
\link[s2bak]{predict.s2bak.so} is a wrapper function that
detects the class of the inputed model and makes the appropriate prediction.

If the provided model is class s2bak.s2 or s2bak.so,
predictions will be made using \link[s2bak]{predict.s2bak.s2}. If an SO
model with BaK is provided (that is, an S2BaK class model with S2 = NULL,
for example through combine without a provided S2), the functions returns
adjusted predictions requiring trait data for the species. If the full
S2BaK model (SO, S2 and BaK are provided), the function will make predictions
with S2 for species with sightings and survey data, and adjusted
predictions using SO and BaK for species without survey data.
}
