% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_etc.R
\name{s2bak.addPred}
\alias{s2bak.addPred}
\title{Adds 'so' (or other specified predictor) to the formula as an additional predictor.}
\usage{
s2bak.addPred(formula, pred = "so")
}
\arguments{
\item{formula}{= Formula or list of formulas to add 'pred' to}

\item{pred}{= Predictor name to add to formula}
}
\value{
A formula with added predictor
}
\description{
Adds 'so' (or other specified predictor) to the formula as an additional predictor. Assumes formula follows a "Y ~ X" format. Does not add 'pred' if it is already in the formula.
}
