% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_funcs.R
\name{s2bak.maxnet}
\alias{s2bak.maxnet}
\title{Modified maxnet function for s2bak}
\usage{
s2bak.maxnet(formula, data, ...)
}
\arguments{
\item{formula}{Formula for fitting the maxnet model}

\item{data}{Fitting dataset including response variable}
}
\value{
Output of the maxnet model
}
\description{
Wrapper function for maxnet::maxnet function. The function
converts the arguments for maxnet to the necessary structure for fitting an
SDM within S2BaK, where the first and second arguments are the formula and
dataset, respectively.
}
