% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_etc.R
\name{s2bak.maxnet}
\alias{s2bak.maxnet}
\title{Wrapper function for maxnet's maxnet function}
\usage{
s2bak.maxnet(formula, data, ...)
}
\arguments{
\item{formula}{= Formula for fitting the maxnet model}

\item{data}{= Fitting dataset including response variable}
}
\value{
Output of the maxnet model
}
\description{
Converts the arguments for maxnet to the necessary structure for fitting an SDM (an 'S3'-like approach)
}
