% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_funcs.R
\name{s2bak.predict}
\alias{s2bak.predict}
\title{Predict with S2BaK objects

Wrapper function that detects the class of the inputed model and makes the appropriate prediction.

If S2 or SO, will make prediction using s2bak.predict.SOS2. If given SO+BaK (that is, an S2BaK class model with S2 = NULL, for example through combine without S2), will make prediction and adjustment which requires trait data for the relevant species. If the full S2BaK model, will make predictions for species in S2, and SO+BaK for the remainder}
\usage{
s2bak.predict(
  model,
  newdata,
  trait = NA,
  predict.fun = predict.gam,
  verbose = FALSE,
  ncores = 1,
  doReadout = FALSE,
  ...
)
}
\arguments{
\item{model}{= Outputted model from s2bak.SO, s2bak.S2, s2bak.S2BaK or s2bak.combine.}

\item{newdata}{= Data to predict}

\item{trait}{= Trait data, optional if model is S2 or SO}

\item{predict.fun}{= prediction function for SDM (for S2 and SO)}

\item{verbose}{= add print statements}

\item{ncores}{= Paralellise SO/S2 predictions, if NA then auto-pick it}

\item{doReadout}{= Whether to force doReadout in SOS2}

\item{...}{= Any inputs for the predict.fun}
}
\value{
Model predictions as a data.frame with columns for each species and rows for each location
}
\description{
Predict with S2BaK objects

Wrapper function that detects the class of the inputed model and makes the appropriate prediction.

If S2 or SO, will make prediction using s2bak.predict.SOS2. If given SO+BaK (that is, an S2BaK class model with S2 = NULL, for example through combine without S2), will make prediction and adjustment which requires trait data for the relevant species. If the full S2BaK model, will make predictions for species in S2, and SO+BaK for the remainder
}
