% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_funcs.R
\name{s2bak}
\alias{s2bak}
\alias{s2bak.S2}
\alias{s2bak.SO}
\alias{s2bak.BaK}
\alias{s2bak.S2BaK}
\title{Build sightings-only or S2 species distribution models for multiple species.

s2bak.SO function fits SDMs for each provided species within the same system, using a specified SDM approach (or the default which are GAMs from the mgcv package). Parallelisation is possible when processing each SDM, with the default being 1 core.

The s2bak.S2 function fits SDMs using species sightings, background sites and survey sites, differentiating between them using a binary 'so' predictor, denoting sightings-only (1) or survey (0).

Saving SDMs to the output may be computationally intensive, particularly with large datasets and many species. To reduce the possibility of crashing, readout and the version = "short" may be used, which does not output the fitted models but instead saves it to the directory specified in readout.}
\usage{
s2bak.S2(
  formula,
  data,
  obs,
  surv = NA,
  background = NA,
  sdm.fun = gam,
  overlapbackground = TRUE,
  nbackground = 10000,
  surv.formula = TRUE,
  ncores = 1,
  verbose = FALSE,
  readout = NA,
  version = c("full", "short")[1],
  ...
)

s2bak.SO(
  formula,
  data,
  obs,
  background = NA,
  sdm.fun = gam,
  overlapbackground = TRUE,
  nbackground = 10000,
  ncores = 1,
  verbose = FALSE,
  readout = NA,
  version = c("full", "short")[1],
  ...
)

s2bak.BaK(predictions, surv, data, trait, verbose = FALSE)

s2bak.S2BaK(
  formula,
  data,
  obs,
  surv,
  trait,
  background = NA,
  sdm.fun = gam,
  predict.fun = predict.gam,
  overlapbackground = TRUE,
  nbackground = 10000,
  surv.formula = TRUE,
  ncores = 1,
  verbose = FALSE,
  readout = NA,
  version = c("full", "short")[1],
  ...
)
}
\arguments{
\item{formula}{Formula for the SDMs, which serves as input for the given SDM. Assumes the structure follows "Y ~ X". Alternatively, a list of formulas can be provided with index names corresponding to species. In this case, species will be fit using their corresponding formula.}

\item{data}{Survey environmental data. Assumes all columns correspond to relevant environmental data and match with predictions/survey data.}

\item{obs}{Species observations as a data.frame, with a column for species name (must be labelled 'species') and column of index of observations to reflect presences. If the index column name is not found in 'data', it assumes row number.}

\item{surv}{Survey data as a matrix, with rows corresponding to environmental data rows.}

\item{background}{Background sites (pseudo-absences) used to fit the presence-only model, provided as a vector of indices of data (following the same column name as observations). If the index column name is not found in 'data', it assumes row number within 'data'. If left as NA, it will randomly sample 'nbackground' sites, with or without overlap ('overlapbackground'). Currently, only one set of background sites can be used.}

\item{sdm.fun}{Model (as function) used for fitting (the default is gam with mgcv). The function must have the formula as their first argument, and 'data' as the parameter for the dataset (including presences and background sites within the data.frame). A wrapper function for maxnet::maxnet() is also available as s2bak.maxnet().}

\item{overlapbackground}{Whether sampled background sites that overlap with observations should be included. By default, it allows overlap. If FALSE, number of background sites may be less than specified or provided.}

\item{nbackground}{Number of background sites to sample. Only applies if background = FALSE.}

\item{surv.formula}{Whether the binary variable 'so', denoting whether sites are sightings-only or survey, should be added to the formula(s). If survey data is not provided, then 'so' will not be added to the formula(s) regardless of whether it is TRUE or FALSE. If there is survey data and surv.formula = FALSE, then 'so' will not be added, and must be included in the initial function call (or it will throw an error).}

\item{ncores}{Number of cores to fit the SDMs, default is 1 core but can be automatically set if ncores=NA. If ncores > number of available cores - 1, set to the latter.}

\item{verbose}{Make print statements}

\item{readout}{Directory to save fitted SDMs and background sites. If NA, it will not save any SDMs. Provides an additional output that shows where the SDM is saved (with file name). The output in this directory can later be used in other functions such as s2bak.predict.SOS2.}

\item{version}{Whether the SDMs should be included in the output. With "short", no the fitted SDMs are not provided. Setting to "full" (default) will output the list with all SDMs. Setting to "short" and combined with readout, can considerably reduce RAM usage while saving the progress so far, which is useful when dealing with many species or large datasets.}

\item{...}{Other arguments that are passed to the SDM function (sdm.fun).}

\item{predictions}{Sightings-only (SO) model predictions over the survey sites for all species, beyond those found in the survey data, as a matrix with columns for each species}

\item{trait}{Full trait data, as a data.frame with 'species' as a column and relevant traits for the rest of them. Do not necessarily have to be involved in the survey, but will be used in the final adjustment model as final output.}
}
\value{
A list of fitted SDMs for each species.

SO class model

Bias adjustment models, the kernels (location and species), as a second-order GLM.

An S2BaK class object containing S2, SO and BaK.
}
\description{
Build S2BaK from top to bottom. Has functionality for parallelisation, but the default is 1 core.
}
\details{
Fits SO models for all species, S2 models for species with survey data. \link{Combine with S2 and SO, note that we include predict.fun, traite and force surv as a mandatory input}
Right now, assumes that all columns/variables in 'data' are relevant for the location bias model (minus the index or whatever).
}
