% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2bak_funcs.R
\name{s2bak}
\alias{s2bak}
\alias{s2bak.S2}
\alias{s2bak.SO}
\title{Build sightings-only or S2 species distribution models for multiple species.

s2bak.SO function fits SDMs for each provided species within the same system, using a specified SDM approach (or the default which are GAMs from the mgcv package). Parallelisation is possible when processing each SDM, with the default being 1 core.

The s2bak.S2 function fits SDMs using species sightings, background sites and survey sites, differentiating between them using a binary 'so' predictor, denoting sightings-only (1) or survey (0).

Saving SDMs to the output may be computationally intensive, particularly with large datasets and many species. To reduce the possibility of crashing, readout and the version = "short" may be used, which does not output the fitted models but instead saves it to the directory specified in readout.}
\usage{
s2bak.S2(
  formula,
  data,
  obs,
  surv = NA,
  background = NA,
  sdm.fun = gam,
  overlapbackground = TRUE,
  nbackground = 10000,
  surv.formula = TRUE,
  ncores = 1,
  verbose = FALSE,
  readout = NA,
  version = c("full", "short")[1],
  ...
)

s2bak.SO(
  formula,
  data,
  obs,
  background = NA,
  sdm.fun = gam,
  overlapbackground = TRUE,
  nbackground = 10000,
  ncores = 1,
  verbose = FALSE,
  readout = NA,
  version = c("full", "short")[1],
  ...
)
}
\arguments{
\item{formula}{= Formula for the SDMs, which serves as input for the given SDM. Assumes the structure follows "Y ~ X". Alternatively, a list of formulas can be provided with index names corresponding to species. In this case, species will be fit using their corresponding formula.}

\item{data}{= Full environmental data used for fitting.}

\item{obs}{= Species observations as a data.frame, with a column for species name (must be labelled 'species') and column of index of observations to reflect presences. If the index column name is not found in 'data', it assumes row number.}

\item{surv}{= Survey data, given as a data.frame of row/index as the first column for the data following the same name as obs, and each column representing a given species presence (1) or absence (0). It will add the an additional binary predictor to the formula(s), "so", which denotes sightings-only or not. If 'so' is already in formula (e.g. if modifying the variable in any way, then set surv.formula = FALSE). If left as NA, it will fit the SDMs as presence-only models with the function of choice. NOTE: Assumes all species are provided in the survey data, and that they are surveyed over the same sites (i.e. matrix-type).}

\item{background}{= Background sites (pseudo-absences) used to fit the presence-only model, provided as a vector of indices of data (following the same column name as observations). If the index column name is not found in 'data', it assumes row number within 'data'. If left as NA, it will randomly sample 'nbackground' sites, with or without overlap ('overlapbackground'). Currently, only one set of background sites can be used.}

\item{sdm.fun}{= Model (as function) used for fitting (the default is gam with mgcv). The function must have the formula as their first argument, and 'data' as the parameter for the dataset (including presences and background sites within the data.frame). A wrapper function for maxnet::maxnet() is also available as s2bak.maxnet().}

\item{overlapbackground}{= Whether sampled background sites that overlap with observations should be included. By default, it allows overlap. If FALSE, number of background sites may be less than specified or provided.}

\item{nbackground}{= Number of background sites to sample. Only applies if background = FALSE.}

\item{surv.formula}{= Whether the binary variable 'so', denoting whether sites are sightings-only or survey, should be added to the formula(s). If survey data is not provided, then 'so' will not be added to the formula(s) regardless of whether it is TRUE or FALSE. If there is survey data and surv.formula = FALSE, then 'so' will not be added, and must be included in the initial function call (or it will throw an error).}

\item{ncores}{= Number of cores to fit the SDMs, default is 1 core but can be automatically set if ncores=NA. If ncores > number of available cores - 1, set to the latter.}

\item{verbose}{= Prints statements of species being fit as we go.}

\item{readout}{= Directory to save fitted SDMs and background sites. If NA, it will not save any SDMs. Provides an additional output that shows where the SDM is saved (with file name). The output in this directory can later be used in other functions such as s2bak.predict.SOS2.}

\item{version}{= Whether the SDMs should be included in the output. With "short", no the fitted SDMs are not provided. Setting to "full" (default) will output the list with all SDMs. Setting to "short" and combined with readout, can considerably reduce RAM usage while saving the progress so far, which is useful when dealing with many species or large datasets.}

\item{...}{= Other arguments that are passed to the SDM function (sdm.fun).}
}
\value{
A list of fitted SDMs for each species.

SO class model
}
\description{
Build sightings-only or S2 species distribution models for multiple species.

s2bak.SO function fits SDMs for each provided species within the same system, using a specified SDM approach (or the default which are GAMs from the mgcv package). Parallelisation is possible when processing each SDM, with the default being 1 core.

The s2bak.S2 function fits SDMs using species sightings, background sites and survey sites, differentiating between them using a binary 'so' predictor, denoting sightings-only (1) or survey (0).

Saving SDMs to the output may be computationally intensive, particularly with large datasets and many species. To reduce the possibility of crashing, readout and the version = "short" may be used, which does not output the fitted models but instead saves it to the directory specified in readout.

Build sightings-only (SO) model, which calls s2bak.S2() but without survey sites. Has functionality for parallelisation, but the default is 1 core.
}
